requirements.txt
Django==1.7.7
cx-Oracle==5.1.3
django-tables2==0.15.0
djangorestframework==2.3.6
psycopg2==2.5.4
pyxml==0.8.4
requests==2.4.1
six==1.8.0
suds==0.4
virtualenv==1.11.6
zsi==2.0-rc3